+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed some logging around expiring games.
+
 ## [v9.6.0](https://github.com/lexicalunit/spellbot/releases/tag/v9.6.0) - 2024-01-30
 
 ### Added